+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,14 @@
 # ui
 
+## 1.11.4
+
+### Patch Changes
+
+- Message shape fix
+- Updated dependencies
+  - @finalyst/runtime-client-gql@1.11.4
+  - @finalyst/shared@1.11.4
+
 ## 1.11.3
 
 ### Patch Changes