+++ CopilotKit/packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @finalyst/shared
 
+## 1.11.4
+
+### Patch Changes
+
+- Message shape fix
+
 ## 1.11.3
 
 ### Patch Changes